Fixes #182 duplicate entries on edit
Also fixed a whole bunch of little things, including deleting an image, thumbnail preview. Currently the code is set up such that an entry can only have one image though this can be changed if necessary. Cleaned up a lot of the ajax stuff.  Still work to be done on the non ajax edit.