Fixes #130 Add invite functionality to hub
Not so much adding invite functionality, rather users can now send a ‘request to join’ the admin of a closed hub and the admin when they log in get a list of requests which they can accept or decline. The user is emailed if they are accpeted. Also if hub is open, you can see the browse page but you cannot see member or entry details.

Really could do with a code review.